refactor: Update authentication resolvers and middleware to include user role
Add a new mutation `changeUserRole` to update the role of a user. This mutation requires the user to be authenticated and have the role of "SUPER_ADMIN". It takes an email and a role as arguments and updates the user's role in the database. If the user is not found, an error is thrown. This change enhances the authentication functionality by allowing authorized users to change the role of other users.